feature: serialize JSON LabelSet as array of strings
This commit improves the JSON serialization of the LabelSet struct by
changing the way labels are represented in JSON. Instead of using a
map with string keys and string values, the labels are now represented
as an array of strings. This change simplifies the JSON structure and
makes it easier to work with labels in a more intuitive way.