feat: adds debug logs to ethclient direct calls (#43)
This PR extends the MultiClient implementation with additional method
overrides to help surface errors originating from the underlying
ethclient. The goal is to improve visibility into the flakiness observed
when running changesets through CI and GAP, and to help identify the
specific call causing the issue.

The additional overrides cover all method calls that can be invoked as
part of a `transact` call, as seen
[here](https://github.com/ethereum/go-ethereum/blob/612c9e0f4ae483f180ef96ace86effec6dc9bb0e/accounts/abi/bind/base.go#L390).