fix(sonar):include lint in schedule main (#104)
This pull request updates the SonarQube workflow in
`.github/workflows/schedule-main.yml` to improve its configuration and
functionality. The changes include removing a conditional check and
adding a new parameter to the SonarQube step.

Updates to SonarQube workflow:

* Removed the conditional `if: github.event_name == 'pull_request'` from
the SonarQube job, allowing it to run regardless of the event type.
* Added a new `include-lint: true` parameter to the SonarQube step to
include linting results in the scan.