[CLD-267]: feat: new blockchain field (#108)
As part of improving the existing design for chains to be more scalable,
we are moving the specific chain code into their own package.

These chains will be abstracted by a higher level field call
BlockChains.

This limits the exposure of Map fields externally and allow the design
to scale better as we introduce more chains to the framework

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-267