fix: Update SUI & Ton to compose ChainMetadata
The original intention was not to compose ChainMetadata and just use it directly by explicitly implementing the BlockChain interface and calling ChainMetadata to avoid breaking existing compatibility with EVM, SOL and APTOS and also the clunky initiation code.

However this introduces some duplication across all chains.

So the compromise is to still have the duplication in the existing chains but for new chains, they will compose the ChainMetadata and to avoid referencing another package for the ChainMetadata, the type alias is introduced.