feat(datastore): Implement Stringer for ContractMetadata and AddressRef Keys (#117)
This pull request introduces enhancements to the `datastore` package by
implementing the `fmt.Stringer` interface for key types and adding
corresponding unit tests. These changes improve the readability and
debugging of key representations.

### Enhancements to key types:

*
[`datastore/address_ref_key.go`](diffhunk://#diff-a8f748a3f09f8874d72a162c5254eff51b1666cedf9059510125a0c4daf5f2e3R4-R13):
Added the `fmt.Stringer` interface to the `AddressRefKey` type,
including a new `String` method that provides a formatted string
representation of the key.
[[1]](diffhunk://#diff-a8f748a3f09f8874d72a162c5254eff51b1666cedf9059510125a0c4daf5f2e3R4-R13)
[[2]](diffhunk://#diff-a8f748a3f09f8874d72a162c5254eff51b1666cedf9059510125a0c4daf5f2e3R60-R69)
*
[`datastore/contract_metadata_key.go`](diffhunk://#diff-d9574b8aedc2795d729a5c1d72ecf6974755c0cb3b4642b28a74e370e9d5a150R3-R9):
Added the `fmt.Stringer` interface to the `ContractMetadataKey` type,
including a new `String` method that provides a formatted string
representation of the key.
[[1]](diffhunk://#diff-d9574b8aedc2795d729a5c1d72ecf6974755c0cb3b4642b28a74e370e9d5a150R3-R9)
[[2]](diffhunk://#diff-d9574b8aedc2795d729a5c1d72ecf6974755c0cb3b4642b28a74e370e9d5a150R39-R43)