feat(datastore): removes EnvMetadataKey implementation (#118)
This PR removes the `EnvMetadataKey` implementation as well as the
`Domain` and `Environment` fields from the `EnvMetadata` struct. These
fields were redundant, given that the `Datastore` can only hold a single
`EnvMetadata` instance. From the perspective of a changeset, the domain
and environment are already implicitly known, making these fields
unnecessary.

This change also simplifies the API by removing the need to specify the
domain and environment each time a user sets `EnvMetadata`.

from 
```
if err = dataStore.EnvMetadata().Set(
		ds.EnvMetadata[exemplarMd.SimpleEnv]{
			Domain:      "exemplar",
			Environment: e.Name,
			Metadata:    envMetadata.Metadata,
		},
	); err != nil {
		return deployment.ChangesetOutput{},
			fmt.Errorf("failed to save updated env metadata in datastore: %w", err)
	}
```

to 

```
if err = dataStore.EnvMetadata().Set(
		ds.EnvMetadata[exemplarMd.SimpleEnv]{
			Metadata:    envMetadata.Metadata,
		},
	); err != nil {
		return deployment.ChangesetOutput{},
			fmt.Errorf("failed to save updated env metadata in datastore: %w", err)
	}
```