feat(operations): allow force run
There is a scenario discovered when an operation with same input has to be run multiple times, eg when deploying the same MCMS contract so it can be associated with different roles.

This commit introduces a new functional option for executeOperation that allow users to force run regardless of previous successful execution.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-302