fix: Aptos Chain Providers now return a value instead of the pointer (#137)
The provider was returning a pointer `*aptos.Chain` which satisfies the
`Blockchain` interface return value. However when calling the
`AptosChains` method, this does a type assertion against the
`aptos.Chain` value type. This resulted in the `AptosChains` method
returning 0 chains.

This commit changes the return value of the `RPCChainProvider` and
`CTFChainProvider` to return the `aptos.Chain` value instead of a
pointer.