refactor: use chain sel from loop instead of chain.ChainSelector() (#146)
No behaviour change, i had to do this because in Chainlink, there are
many tests where they do
```
 cldf_chain.NewBlockChains(map[uint64]cldf_chain.BlockChain{
		chainselectors.SOLANA_DEVNET.Selector: cldf_solana.Chain{},
	})
```
Discover [in this PR with the CI
failures](https://github.com/smartcontractkit/chainlink-deployments/pull/3548)

So when we do .EVMChains(), the chainselector is 0 because
cldf_solana.Chain{} is set to default value.

I could update the code on Chainlink but this involves more code change.