feat(operations): introduce ExecutionOperationN
There is a scenario discovered when an operation with same input has to be run multiple times, eg when deploying the same MCMS contract so it can be associated with different roles.

This commit introduces ExecutionOperationN, will execute the provided operation N times.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-302