feat(datastore) remove generics from the top level api (#149)
This PR introduces a leaner custom metadata handling model by replacing
generics with `any`. While generics are still used to define interfaces,
we now replace the use of generics for contract and environment metadata
with any. This simplifies the API and removes the need for two
conversion steps per changeset. Users can continue to work with their
own metadata types. The only trade-off is that when reading from the
datastore (e.g., after a Get()), they’ll need to convert from any, which
is easily handled using the provided `As` helper. Overall, this change
reduces boilerplate and improves usability without sacrificing
flexibility.