[CLD-302]: feat(operations): introduce ExecutionOperationN (#152)
There is a scenario discovered when an operation with same input has to
be run multiple times, eg when deploying the same MCMS contract multiple
times so it can be associated with different roles.

This commit introduces `ExecutionOperationN`, it will execute the
provided operation N times.

For idempotency behaviour, we went to ensure on reruns, the
`ExecutionOperationN` will return the successful prior results in the
same order in the slice. That is why we need to remember the order.

Usage
```go
// execute 3 times
linkDeployReports, err := operations.ExecuteOperationN(
			b, DeployLinkOp, deps, operations.EmptyInput{},
			"deploy-multiple-link-contract", 3,
			operations.WithRetry[operations.EmptyInput, EthereumDeps](),
		)
```

Report
```
  {
    "id": "15ac8420-0b59-4293-a188-948ad43643da",
    "definition": {
      "id": "deploy-link-token-op",
      "version": "1.0.0",
      "description": "Deploy LINK Contract Operation"
    },
    "output": "0x9dc98d0a1362d1aa83a10569c45ef4ee3ef97071",
    "input": {},
    "timestamp": "2025-06-04T15:36:02.066618+10:00",
    "error": null,
    "childOperationReports": null,
    "executionSeries": {
      "id": "deploy-multiple-times",
      "order": 1
    }
  }
```

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-302