feat(operation): support execution of dynamic operation
To enhance the flexibility of Operations API, we want to support dynamic execution of Operation.
Imagine this snippet of code:

```
    inputs := []any{"input1", 42}
	deps := []any{Deps1{}, Deps2{}}
	defs := []Definition{
		{ID: "string-op", Version: semver.MustParse("1.0.0")},
		{ID: "int-op", Version: semver.MustParse("1.0.0")},
	}

	// dynamically retrieve and execute operations on different inputs
	for i, def := range defs {
	    // registry contains a list of operations
		retrievedOp, _ := registry.Retrieve(def)
		// dependency and input are dynamic/different per Operation
		report, _ := ExecuteOperation(b, retrievedOp, deps[i], inputs[i])
	}
```

Based on the definition, we can decide on run time what operations to execute and what input and deps to pass in to that operation. This can be quite powerful to allow users to dynamically construct a series of operation under a sequence.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-354