[CLD-486]: feat: new CTF Provider for JD (#241)
Introduce new CTF provider for JD which provides a ready for testing JD
docker container.

Because JD docker image is not available publicly, we need to use ECR
auth to gain access to the docker image, the ctf_integration_test.go
which spins up the JD containers and execute requests on it is currently
skipped because of it. Or just build it locally.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-486