ci: optimize go test config (#267)
- set timeout for test to 10 min so it does not hang indefinitely.
- set go test config -p 2 -parallelism 3 which means at any time run 3
tests over max of 2 different packages. This limits the number of docker
containers running at any time.
 
I did a bunch of testing and find this give the best stablilty and
performance without going for bigger runner such as
`ubuntu-latest-8-cores`.

With `ubuntu-latest` and `-p 1 -parallelism 1` , it works without
hanging, since everything is serialized but that takes a long time to
run. However with `-p 2 -parallelism 3` which means max of 6 containers
at anytime seem to work nicely.

This test with new config still runs under 5 minutes and we no longer
have the test hanging for ages.