feat: Replace WithRPCURLTransformer with targeted transformers (#270)
The previous `WithRPCURLTransformer` load option was a single function
that transformed both HTTP and WS URLs. This is too general and makes
the assumption that the caller wants to transform all the URLs with the
same function.

This PR introduces 2 new load options: `WithHTTPURLTransformer` and
`WithWSURLTransformer` which are more specific and allow the caller to
transform only or both the HTTP or WS URLs, with different functions.