[CLD-518]: feat: new anvil evm provider (#298)
Introducing a new anvil CTF provider for EVM, this allows us to use
anvil for fork testing.

The reason we are doing this is because in CLD, when loading the forked
environment for mcms fork testing, it spins up anvil and uses a bunch of
stuff from CHainlink/deployment like kms client and devenv.NewChains,
but implementing this framework standard anvil provider, we can replace
everything there.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-518