fix: migrate proposejob & registernode
Migrate Propose Job and Register Node from pkg/offchain in CLD to CLDF.

Also added tests as it didn have any.
In order to mock the JD client in engine, i needed to move the mocks in offchain package to a place that is accessible for both offchain and engine package.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-568