feat: implement env config loading for environment
This introduces a Config struct for the environment config, which
contains the network config and the env config. The network and env
config are loaded from the .config directory in the domain.

As part of this move, the networks package has been removed, and the
code has been moved to the environment package to handle an import
cycle.