feat: move engine `environment.Config` to `config.Config`
This PR moves the `environment.Config` to `config.Config` and updates
all the relevant imports to use the new `config.Config`.

This change is designed to improve the clarity and organization of the
configuration code, colocating it with all the other config code.

As part of this change, we have also changed the name of
`environment.LoadConfig` to `config.Load`.

All functionality remains the same.