[CLD-606]: feat(chain): addr string to bytes conversion (#386)
Context:
https://chainlink-core.slack.com/archives/C08QF1BEW4T/p1756934131485399

Design thoughts:
Initially i went with [this
design](https://github.com/smartcontractkit/chainlink-deployments-framework/pull/376)
where i introduced the AddressToBytes method to the blockchain
interface. However i didn quite like this approach because
`AddressToBytes` feels out of place in the blockchain interface (it
could be its own interface, and to convert address, we dont need the
`chain` struct) and this also forces all implementations of blockchain
to implement `AddressToBytes`.

So i went with the approach in this PR, introducing a new interface call
AddressConverter and using the strategy pattern for each chain to
provide the conversion functionality. This decouples the Address
conversion logic from the Blockchain interface.


This commit adds support for converting string address to bytes for each
of the supported chain.

### ToBytes
```
		ethChain := e.BlockChains.EVMChains()[0]
		bytes, err := address.ToBytesForFamily(ethChain.Family(), "0x742d35Cc6634C0532925a3b8D4c8C1B8c4c8C1B8")
```


JIRA: https://smartcontract-it.atlassian.net/browse/CLD-606