feat: add `reset-proposal` command (#410)
Adds a new command to the CLI `reset-proposal` which resets signatures
and fetches the latest opcount on chain. This is useful when proposals
get an opcount related error and need to be resigned. We can use this
command to clear signatures and get latest opcounts without manually
changing the proposal. We will use this on the CLD repo inside the
pipelines via a comment like `/reset-proposal`

---------

Co-authored-by: gustavogama-cll <165679773+gustavogama-cll@users.noreply.github.com>