feat(environment): use network config chains instead of addressbook (#425)
This PR removes the dependency on `addressbook.json` for chain loading
in load environment. Previously, for a chain to be loaded it was
necessary to add an entry for the corresponding chain selector in
`addressbook.json`, even when it wasn't being used. Now it loads all
chain selectors from the network configuration by default.