[CLD-691]: fix: preserve large integers in YAML to JSON conversion (#459)
Raised in the support ticket
-https://chainlink-core.slack.com/archives/C08QF1BEW4T/p1758918706928359

**Problem:**
When we execute pipeline, we convert YAML -> JSON -> Code

This is what is happening when there is a really big integer
- YAML parsing converts large numbers like `2000000000000000000000` to
`float64(2e+21)`
- JSON marshaling converts `float64(2e+21)` to scientific notation
`"2e+21"`
- big.Int cannot unmarshal scientific notation, causing errors

This commit ensure value is converted back to the original format
instead of scientific notation.

Long term fix: because of legacy reason, we convert yaml -> json ->
code, however we if refactor and remove the middle json, it will make
this redundant and cleaner. Something to do in the future.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-691