[CLD-700]: feat: add --changeset-index flag for array format YAML files (#473)
Raised in this support ticket -
https://chainlink-core.slack.com/archives/C08QF1BEW4T/p1759262737294269

We have an issue where if the in the same input yaml, we have multiple
changeset with the same name, the command does not know which one to
execute and will execute the first one all the time.

This bug was introduced when we refactor the CI logic to use the durable
pipeline run cmd for pipeline execution. Since the cmd does not support
running multiple changeset with same name in the same input yaml file,
the bug was created.

Add support for running changesets by index position to handle duplicate
changeset names in array format input files.

Examples:
```
  # Run first changeset by index
  chainlink-deployments durable-pipeline run --environment testnet \
    --input-file inputs.yaml --changeset-index 0

  # Run changeset by name (existing functionality)
  chainlink-deployments durable-pipeline run --environment testnet \
    --input-file inputs.yaml --changeset my_changeset
```

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-700