feat(operations): introduce AsUntypedRelaxed
Add AsUntypedRelaxed method for a more relax type checking relying on JSON round-trip conversion
for YAML-unmarshaled inputs.

This is needed because when input is a struct , when input is unmarshal from yaml into any, the default type is map[string]any which does not match the expected input type for the operation/sequence, allowing the option of marshaling for AsUntyped allow users to force the conversion instead of relying only on type assertion.