fix(catalog/memory): replace pgtest with map
The current memory catalog client implementation uses pgtest which requires postgres binary available in the host machine and also actually spins up a postgres service when running the test. We want a true in memory implementation.

Remove usage of pgtest and replace with map for a true inmemory client implementation of catalog

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-755