feat(pipeline): enable strict yaml unmarshal
Originated from [this support request](https://chainlink-core.slack.com/archives/C09AJMWA98E/p1759973431200199?thread_ts=1759969075.884959&cid=C09AJMWA98E), when user has a typo of the field, the yaml will be unmashaled and that field will be ignored , this can cause errors in the pipeline when a certain expected field is not defined because the field was a typo.

Enable strict unmarshaling and throws an error when there is an unknown field in the yaml input.

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-759