fix: removing chainlink-common dep as direct dependency (#524)
Attempt to remove dependency on chainlink-common, cloning the required
interface for logger and keystore interface
However we still have indirect dependency on chainlink-common from our
other dependencies.

We want to remove dependency on `chainlink-common` because we use it
only for the logger, and `chainlink-common` tend to have breaking
changes with bad versioning.

Chainlink repo run with this commit -
https://github.com/smartcontractkit/chainlink/pull/19967 - everything
compiled normally

JIRA: https://smartcontract-it.atlassian.net/browse/CLD-777