chore: enable dumping the container logs of the forked environment
Dump the container logs when the CLDF_CONTAINER_LOGS environment
variables is set. If the environment variable is set to an absolute
filesystem path, CLDF tries to save the logs to a folder at the
given path. If it's set to a _truish_ value ("true", "on", "enabled",
"1"), it saves the logs to a temporary folder.