feat: add job filter to memory client
When calling `ListJobs`, the filter can now be applied to the jobs.

This supports all the filter options that are supported by the
`ListJobs` method. It also makes a change to the DeleteJob method to
perform a soft delete of the job, rather than a hard delete. This is in
line with how the Job Distributor API works.