fix(catalog): updates errors in remote implementation (#551)
`LoadDataStoreFromCatalog` should not produce an error when fetching
parts of the datastore for a specific domain environment if no records
are returned. The implementation treats `ErrAddressRefNotFound`,
`ErrChainMetadataNotFound`, `ErrContractMetadataNotFound`, and
`ErrEnvMetadataNotSet` as valid cases of an empty datastore. This is
useful when a domain environment has no data during the first execution.

However, the remote catalog implementation fails to return
`ErrAddressRefNotFound`, `ErrChainMetadataNotFound`,
`ErrContractMetadataNotFound`, and `ErrEnvMetadataNotSet` correctly,
causing `LoadDataStoreFromCatalog` to fail when no records are found.