fix(mcms): select public rpc for fork tests (#553)
### What

Ignore CLL rpcs when selecting the RPC for the ForkedEnvironment (used
by anvil to fetch
the contract state).

### Why

In the CI, the fork tests need to GAP to access the CLL rpcs. However,
because we run anvil inside a containers, the setup gets rather
complicated -- we need to copy+adapt the entries added to `/etc/hosts`
and add GAP's custom CA certificates. But, because anvil is managed
using testcontainers (indirectly, via CTF), we'd need to add support for
this feature to CTF, then CLDF and finally the CLD workflows.

However, CLL is moving towards replacing GAP with Tailscale to access
internal resources from GHA runners. So, rather than spending
significant resources implementing the GAP solution, we're adding this
quick fix -- which should solve the problem for 95% of the proposals --
until the Tailscale option is ready.

---
DPT-277