feat(mcms): fetch pipeline PR data before decoding a proposal (#633)
Now that proposal PRs are created targeting the `main` branch, it's
possible that the decoder won't have all the data it needs in the
proposal's own branch. The typical case when this happens is when a
pipeline changeset deploys a contract and then a subsequent changeset
references the new address.

This PR addresses this limitation by fetching the updated address book
(or datastore/address_refs file) when a proposal has the required
metadata indicating what is the pipeline PR that created it.